fix(client): resolve react-refresh/only-export-components lint errors
Extract non-component exports (cva variants, useFormField hook) from
shadcn/ui component files into dedicated modules so each .tsx file
only exports React components, satisfying Fast Refresh requirements.
